% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{analysis_suite_dir}
\alias{analysis_suite_dir}
\title{Return path to root of AnalysisSuite package}
\usage{
analysis_suite_dir(...)
}
\arguments{
\item{...}{Additional arguments (typically path components) passed to file.path}
}
\description{
Return path to root of AnalysisSuite package
}
\seealso{
\code{\link{file.path}}
}

