\name{install_analysis_suite}
\alias{install_analysis_suite}
\title{Download/install AnalysisSuite (or link prior download) to standard location}
\usage{
  install_analysis_suite(path = NULL, ref = "master")
}
\arguments{
  \item{path}{Optional non-standard path that will be
  symlinked to standard locatio (see details)}

  \item{ref}{Git reference ()}
}
\description{
  With default parameters, this function will download the
  current master of AnalysisSuite and install to a standard
  location
}

