% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_bioc}
\alias{install_bioc}
\title{Install any required Biocondutcor packages}
\usage{
install_bioc(pkgs = "RBGL")
}
\arguments{
\item{pkgs}{Character vector of packages}
}
\description{
Install any required Biocondutcor packages
}
\author{
jefferis
}

