% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_from_zip}
\alias{install_from_zip}
\title{Install source code from a zip file to a specified folder inside nat.as package}
\usage{
install_from_zip(zip_file, install_dir)
}
\arguments{
\item{zip_file}{Zip file containing source code to install}

\item{install_dir}{Name of subdirectory that will containing unzipped code}
}
\value{
logical indicating success
}
\description{
Install source code from a zip file to a specified folder inside nat.as package
}
\author{
jefferis
}

