\name{.require_analysis_suite}
\alias{.require_analysis_suite}
\title{nat.as internal functions}
\usage{
  .require_analysis_suite()
}
\description{
  This function (which is designed for internal use only)
  will Start AnalysisSuite by sourcing all function
  definitions and loading key libraries. Users should load
  the code with \code{library(nat.as)} or
  \code{reload_analysis_suite()}.
}
\seealso{
  \code{\link{reload_analysis_suite}}
}

