\docType{package}
\name{nat.as-package}
\alias{nat.as}
\alias{nat.as-package}
\title{Analyse 3D biological image data especially neurons}
\description{
  Key functions that you will want to look at include
  read.neuron write.neuron
}
\details{
  Objects of class neuron neuronlist

  plot.neuron plot3d.neuron
}
\seealso{
  \code{\link[rgl]{rgl-package}} which is used for
  visualisation.
}
\keyword{package}

