% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{url_header}
\alias{url_header}
\title{Get header information from a file}
\usage{
url_header(url)
}
\arguments{
\item{url}{URL to fetch header}
}
\description{
Get header information from a file
}
\seealso{
\code{\link{HEAD}}
}

