\name{install_from_zip}
\alias{install_from_zip}
\title{Install source code from a zip file to a specified folder inside nat.as package}
\usage{
  install_from_zip(zip_file, install_dir = "AnalysisSuite")
}
\arguments{
  \item{zip_file}{Zip file containing source code to
  install}

  \item{install_dir}{Subdirectory of nat.as package}
}
\value{
  logical indicating success
}
\description{
  Install source code from a zip file to a specified folder
  inside nat.as package
}
\author{
  jefferis
}

